@@ -5,7 +5,6 @@ import (
 	"encoding/base64"
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
 	"net/http"
 	"net/url"
 	"os"
@@ -134,7 +133,7 @@ func AppCommand() cli.Command {
 		Action:  defaultAction(appLs),
 		Flags:   appLsFlags,
 		Subcommands: []cli.Command{
-			cli.Command{
+			{
 				Name:        "ls",
 				Usage:       "List apps",
 				Description: "\nList all apps in the current Rancher server",
@@ -142,13 +141,13 @@ func AppCommand() cli.Command {
 				Action:      appLs,
 				Flags:       appLsFlags,
 			},
-			cli.Command{
+			{
 				Name:      "delete",
 				Usage:     "Delete an app",
 				Action:    appDelete,
 				ArgsUsage: "[APP_NAME/APP_ID]",
 			},
-			cli.Command{
+			{
 				Name:        "install",
 				Usage:       "Install an app template",
 				Description: installAppDescription,
@@ -194,7 +193,7 @@ func AppCommand() cli.Command {
 					},
 				},
 			},
-			cli.Command{
+			{
 				Name:      "rollback",
 				Usage:     "Rollback an app to a previous version",
 				Action:    appRollback,
@@ -210,7 +209,7 @@ func AppCommand() cli.Command {
 					},
 				},
 			},
-			cli.Command{
+			{
 				Name:        "upgrade",
 				Usage:       "Upgrade an existing app to a newer version",
 				Description: upgradeAppDescription,
@@ -247,7 +246,7 @@ func AppCommand() cli.Command {
 					},
 				},
 			},
-			cli.Command{
+			{
 				Name:        "list-templates",
 				Aliases:     []string{"lt"},
 				Usage:       "List templates available for installation",
@@ -262,7 +261,7 @@ func AppCommand() cli.Command {
 					},
 				},
 			},
-			cli.Command{
+			{
 				Name:        "show-template",
 				Aliases:     []string{"st"},
 				Usage:       "Show versions available to install for an app template",
@@ -270,7 +269,7 @@ func AppCommand() cli.Command {
 				ArgsUsage:   "[TEMPLATE_ID]",
 				Action:      templateShow,
 			},
-			cli.Command{
+			{
 				Name:      "show-app",
 				Aliases:   []string{"sa"},
 				Usage:     "Show an app's available versions and revisions",
@@ -280,7 +279,7 @@ func AppCommand() cli.Command {
 					formatFlag,
 				},
 			},
-			cli.Command{
+			{
 				Name:      "show-notes",
 				Usage:     "Show contents of apps notes.txt",
 				Action:    appNotes,
@@ -799,7 +798,7 @@ func walkTemplateDirectory(templatePath string) (string, map[string]string, erro
 			return nil
 		}
 		version := &chartVersion{}
-		content, err := ioutil.ReadFile(path)
+		content, err := os.ReadFile(path)
 		if err != nil {
 			return err
 		}
@@ -816,7 +815,7 @@ func walkTemplateDirectory(templatePath string) (string, map[string]string, erro
 			if info.IsDir() {
 				return nil
 			}
-			content, err := ioutil.ReadFile(path)
+			content, err := os.ReadFile(path)
 			if err != nil {
 				return err
 			}
@@ -1254,7 +1253,7 @@ func parseAnswersFile(location string, answers map[string]string) error {
 }
 
 func parseFile(location string) (map[string]interface{}, error) {
-	bytes, err := ioutil.ReadFile(location)
+	bytes, err := os.ReadFile(location)
 	if err != nil {
 		return nil, err
 	}
